+++ resources/views/default/display/tree.blade.php
@@ -1,7 +1,7 @@
 @yield('before.card')
 @yield('before.panel')
 
-<div class="card card-default {!! $panel_class !!}">
+<div class="card card-default {!! $card_class !!}">
     <div class="card-heading card-header">
         @if ($creatable)
             <a class="btn btn-primary mt-2" href="{{ $createUrl }}">