+++ resources/views/default/form/card.blade.php
@@ -5,7 +5,7 @@
   <input type="hidden" name="_method" value="post" />
   <input type="hidden" name="_redirectBack" value="{{ $backUrl }}" />
   <input type="hidden" name="_token" value="{{ csrf_token() }}" />
-
+  
   {!! $buttons->render() !!}
 
 </form>