+++ resources/views/default/form/panel/element.blade.php
@@ -1,3 +1,0 @@
-<div {!! $attributes !!}>
-    @include(AdminTemplate::getViewPath('form.partials.elements'), ['items' => $elements])
-</div>