+++ src/Display/Column/Email.php
@@ -10,11 +10,6 @@ class Email extends NamedColumn
     protected $view = 'column.email';
 
     /**
-     * @var bool
-     */
-    protected $isSearchable = true;
-
-    /**
      * @return array
      */
     public function toArray()