+++ src/Display/Column/Image.php
@@ -22,7 +22,7 @@ class Image extends NamedColumn
     /**
      * @var bool
      */
-    protected $isSearchable = true;
+    protected $isSearchable = false;
 
     /**
      * @var string