+++ src/Display/Column/Link.php
@@ -10,11 +10,6 @@ class Link extends Url
     protected $view = 'column.link';
 
     /**
-     * @var bool
-     */
-    protected $isSearchable = true;
-
-    /**
      * Check if instance editable.
      *
      * @return bool